+++ src/pysight/tkinter_gui_multiscaler.py
@@ -587,22 +587,6 @@ class GUIApp(object):
                     self.config = {}
             self.__modify_vars()
 
-    @property
-    def outputs(self):
-        """
-        Create a dictionary with the wanted user outputs.
-        """
-        output = {}
-
-        if True == self.summed.get(): output['summed'] = True
-        if True == self.memory.get(): output['memory'] = True
-        if True == self.stack.get(): output['stack'] = True
-
-        if 'stack' in output:
-            if not 'summed' in output and not 'memory' in output:
-                warnings.warn("Performance Warning: Writing data to file might take a long time when the required"
-                              " output is only 'Full Stack'.")
-        return output
 
 def verify_gui_input(gui):
     """Validate all GUI inputs"""